create endpoints and separate files to allow for api testing
